+++ components/common/nav/types/nav.ts
@@ -1,0 +1,5 @@
+export interface NavListProps {
+  href: string;
+  text: string;
+  SvgValue: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+}
